Made EventWrapper class abstract
