+++ src/utils/EventHandler.ts
@@ -1,12 +1,10 @@
-import {EventEmitter} from "events";
-
-export class EventHandler {
-	private readonly _wrappedName: string;
+export abstract class EventHandler {
+	protected readonly _wrappedName: string;
 	private readonly _apiEventName: string;
 	private readonly _events: object;
 	private readonly _isInitEvent: boolean;
 
-	constructor(name: string, events: object) {
+	protected constructor(name: string, events: object) {
 		this._events = events;
 		if (events.hasOwnProperty(name)) {
 			this._apiEventName = events[name].name;
@@ -31,16 +29,5 @@ export class EventHandler {
 		} else {
 			return null;
 		}
-	}
-
-	wrap(originalEmitter: any, wrappedEmitter: EventEmitter): void {
-		originalEmitter.on(this.getApiEventName(), (object) => {
-			let WrappedObject = this.getWrappedObject(object);
-			if (WrappedObject) {
-				wrappedEmitter.emit(this._wrappedName, WrappedObject);
-			} else {
-				wrappedEmitter.emit(this._wrappedName);
-			}
-		});
 	}
 }
