+++ src/utils/EventWrapper.ts
@@ -1,5 +1,5 @@
-import {EventHandler} from "./EventHandler";
 import {EventEmitter} from "events";
+import {DiscordEventHandler} from "../wrappers/discord/DiscordEventHandler";
 
 export class EventWrapper {
 	private readonly _originalEmitter: any;
@@ -13,7 +13,7 @@ export class EventWrapper {
 	registerEvents(eventsToRegister: object, excludeInitEvents?: boolean) {
 		for (let name in eventsToRegister) {
 			if (eventsToRegister.hasOwnProperty(name)) {
-				let Event: EventHandler = new EventHandler(name, eventsToRegister);
+				let Event: DiscordEventHandler = new DiscordEventHandler(name, eventsToRegister);
 				let excludeEvent = excludeInitEvents ? Event.isInitEvent() : false;
 
 				if (!excludeEvent) {
